core: move login5 into own manager
